feat: configurable diagonal movement (VI)
* maintainer's review as 03/03/2022 (WIP)

+++ Intersect.Client/Entities/Dash.cs
@@ -10,7 +10,7 @@ namespace Intersect.Client.Entities
     public partial class Dash : IDash
     {
 
-        private sbyte mChangeDirection = -1;
+        private int mChangeDirection = -1;
 
         private int mDashTime;
 
@@ -34,7 +34,7 @@ namespace Intersect.Client.Entities
 
         public float OffsetY => GetYOffset();
 
-        public Dash(Entity en, Guid endMapId, byte endX, byte endY, int dashTime, sbyte changeDirection = (sbyte)Direction.None)
+        public Dash(Entity en, Guid endMapId, byte endX, byte endY, int dashTime, int changeDirection = (int)Direction.None)
         {
             mChangeDirection = changeDirection;
             mEndMapId = endMapId;
